adds _IFS_collection.zip and lengths for URLs

+++ src/main/java/org/iupac/fairspec/api/IFSSerializerI.java
@@ -8,7 +8,7 @@ public interface IFSSerializerI {
 
 	void addAttr(String key, String val);
 
-	void addAttrInt(String key, int ival);
+	void addAttrInt(String key, long ival);
 
 	void addValue(Object val);
